working state pt 2
