+++ apps/modernization-api/src/main/java/gov/cdc/nbs/config/NBSEntityConfiguration.java
@@ -9,8 +9,6 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
  * the database-entities project.
  */
 @Configuration
-@EnableJpaRepositories(
-    basePackages = "gov.cdc.nbs.repository",
-    repositoryBaseClass = gov.cdc.nbs.repository.detach.DetachableJpaRepository.class)
+@EnableJpaRepositories(basePackages = "gov.cdc.nbs.repository")
 @EntityScan("gov.cdc.nbs.entity")
 class NBSEntityConfiguration {}
